feat(Tabs)!: combine Tab sub-component into Tabs
- update imports, tests, and usages in Tabs
- update usage in WireframeDemo
- clean up snapshots
