--- src/components/Tab/index.ts
@@ -1,2 +1,0 @@
-export { Tab as default } from './Tab';
-export type { TabProps } from './Tab';
